split Weather component for better maintenance
