+++ src/components/weather/Weather.comp.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { WeatherHeading } from './components/WeatherHeading.comp'
-import { WeatherProvider } from './hooks/useWeather'
-import { WeatherLocation } from './components/WeatherLocation.comp.jsx'
+import { WeatherProvider } from './components/WeatherMain/hooks/useWeather'
+import { WeatherMain } from './components/WeatherMain/WeatherMain'
 import { Flex } from '@chakra-ui/react'
 
 export const WeatherComp = () => {
@@ -9,7 +9,7 @@ export const WeatherComp = () => {
     <WeatherProvider>
       <Flex flexDirection={'column'} alignItems={'center'} gap={'1rem'}>
         <WeatherHeading/>
-        <WeatherLocation/>
+        <WeatherMain/>
       </Flex>
     </WeatherProvider>
   )
